requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Schema validation
jsonschema==4.20.0
fastavro==1.9.0
avro-python3==1.10.2

# Streaming platforms
kafka-python==2.0.2
confluent-kafka==2.3.0
aiokafka==0.10.0

# AWS (optional)
boto3==1.34.0
botocore==1.34.0

# GCP (optional)
google-cloud-pubsub==2.18.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Storage
minio==7.2.0

# API & Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.0

# Monitoring
prometheus-client==0.19.0
statsd==4.0.1

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
click==8.1.7
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Async
asyncio==3.4.3
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==21.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
